move flutter_theoplayer_sdk files under src
