+++ flutter_theoplayer_sdk/lib/src/theoplayer_internal.dart
@@ -4,9 +4,9 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:theoplayer/widget/chromeless_widget.dart';
-import 'package:theoplayer/widget/fullscreen_widget.dart';
-import 'package:theoplayer/widget/presentationmode_aware_widget.dart';
+import 'package:theoplayer/src/widget/chromeless_widget.dart';
+import 'package:theoplayer/src/widget/fullscreen_widget.dart';
+import 'package:theoplayer/src/widget/presentationmode_aware_widget.dart';
 import 'package:theoplayer_platform_interface/helpers/logger.dart';
 import 'package:theoplayer_platform_interface/pigeon/apis.g.dart';
 import 'package:theoplayer_platform_interface/platform/platform_activity_service.dart';
@@ -17,9 +17,9 @@ import 'package:theoplayer_platform_interface/theoplayer_view_controller_interfa
 import 'package:theoplayer_platform_interface/track/mediatrack/theoplayer_audiotrack.dart';
 import 'package:theoplayer_platform_interface/track/mediatrack/theoplayer_videotrack.dart';
 import 'package:theoplayer_platform_interface/track/texttrack/theoplayer_texttrack.dart';
-import 'package:theoplayer/theoplayer_state.dart';
+import 'package:theoplayer/src/theoplayer_state.dart';
 import 'package:theoplayer/theoplayer_tracklist_wrapper.dart';
-import 'package:theoplayer/theoplayer_view.dart';
+import 'package:theoplayer/src/theoplayer_view.dart';
 
 // TODO: check this, quick fix for the main.dart
 // TODO: move exports into a separate file
@@ -33,8 +33,8 @@ export 'package:theoplayer_platform_interface/track/mediatrack/theoplayer_videot
 export 'package:theoplayer_platform_interface/track/mediatrack/theoplayer_videotrack_events.dart';
 export 'package:theoplayer_platform_interface/track/texttrack/theoplayer_texttrack_events.dart';
 export 'package:theoplayer_platform_interface/track/texttrack/theoplayer_texttrack.dart';
-export 'package:theoplayer/widget/chromeless_widget.dart';
-export 'package:theoplayer/widget/fullscreen_widget.dart';
+export 'package:theoplayer/src/widget/chromeless_widget.dart';
+export 'package:theoplayer/src/widget/fullscreen_widget.dart';
 
 /// Callback that's used at player creation.
 /// This is called once the underlying native THEOplayerView is fully created and available to use.
