+++ flutter_theoplayer_sdk/lib/src/theoplayer_state.dart
@@ -1,5 +1,5 @@
 import 'package:theoplayer_platform_interface/theoplayer_event_manager.dart';
-import 'package:theoplayer/theoplayer_internal.dart';
+import 'package:theoplayer/src/theoplayer_internal.dart';
 
 //TODO: move this into theoplayer.dart
