+++ flutter_theoplayer_sdk/lib/src/widget/chromeless_widget.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:theoplayer/theoplayer_internal.dart';
-import 'package:theoplayer/widget/presentationmode_aware_widget.dart';
+import 'package:theoplayer/src/theoplayer_internal.dart';
+import 'package:theoplayer/src/widget/presentationmode_aware_widget.dart';
 
 class ChromelessPlayerView extends StatefulWidget {
   const ChromelessPlayerView({
