+++ flutter_theoplayer_sdk/lib/src/widget/fullscreen_widget.dart
@@ -4,8 +4,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-import 'package:theoplayer/theoplayer_internal.dart';
-import 'package:theoplayer/widget/presentationmode_aware_widget.dart';
+import 'package:theoplayer/src/theoplayer_internal.dart';
+import 'package:theoplayer/src/widget/presentationmode_aware_widget.dart';
 
 class FullscreenStatefulWidget extends StatefulWidget {
   final THEOplayer theoplayer;
